pyproject.toml
[tool.poetry]
name = "ml-training"
version = "0.1.0"
description = ""
authors = ["Masato Naka <masatonaka1989@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = ">=3.11,<3.13"



[tool.poetry.group.dev.dependencies]
ruff = "^0.6.0"


[tool.poetry.group.ray.dependencies]
pyarrow = "17.0.0"
pandas = "2.2.3"
ray = {extras = ["train"], version = "^2.10.0"}


[tool.poetry.group.river.dependencies]
river = "0.21.2"
flask = "2.3.3"


[tool.poetry.group.tensorflow.dependencies]
portpicker = "1.6.0"
tensorflow = "2.17.0"
tf-keras = "^2.17.0"
tensorflow-datasets = "^4.9.6"
tensorflow-recommenders = "0.7.2"


[tool.poetry.group.pytorch.dependencies]
torchvision = "^0.19.0"
torch = "^2.2.2"


[tool.poetry.group.streamlit.dependencies]
streamlit = "^1.39.0"
google-cloud-aiplatform = "^1.70.0"
google-auth = "^2.35.0"
google-cloud-run = "^0.10.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
